Move DatabaseHandler into Storage class
