+++ storage/CMakeLists.txt
@@ -6,8 +6,6 @@ set(SOURCES
     include/Storage.hpp
     include/Item.hpp
     src/Storage.cpp
-    src/DatabaseHandler.hpp
-    src/DatabaseHandler.cpp
     src/Timer.hpp
 )
